Fixed errors in code that for some reason were in the main branch of GitHub but worked fine on my end. Despite it being after the submission period, I hope this will be considered

+++ SparseNode.cpp
@@ -36,7 +36,7 @@ int SparseNode::getRow() {
 int SparseNode::getCol() {
     return col;
 }
-int SparseNode::getValue() {
+double SparseNode::getValue() {
     return value;
 }
